+++ filters/base_filter.cpp
@@ -9,16 +9,16 @@ void BaseFilterWithMatrix::Apply(Image& image) {
     }
 
     Image filtered_image = image;
-    const long long y_size = image.pixels.size();
-    const long long x_size = image.pixels[0].size(); 
+    const int64_t y_size = static_cast<int64_t>(image.pixels.size());
+    const int64_t x_size = static_cast<int64_t>(image.pixels[0].size());
 
-    for (long long i = 0; i < y_size; ++i) {
-        for (long long j = 0; j < x_size; ++j) {
+    for (int64_t i = 0; i < y_size; ++i) {
+        for (int64_t j = 0; j < x_size; ++j) {
             Pixel new_pixel(0.0, 0.0, 0.0);
-            for (long long shift_i = -1; shift_i <= 1; ++shift_i) {
-                for (long long shift_j = -1; shift_j <= 1; ++shift_j) {
-                    long long actual_used_pixel_i = shift_i + i;
-                    long long actual_used_pixel_j = shift_j + j;
+            for (int64_t shift_i = -1; shift_i <= 1; ++shift_i) {
+                for (int64_t shift_j = -1; shift_j <= 1; ++shift_j) {
+                    int64_t actual_used_pixel_i = shift_i + i;
+                    int64_t actual_used_pixel_j = shift_j + j;
 
                     if (0 > actual_used_pixel_j) {
                         actual_used_pixel_j = 0;
@@ -34,9 +34,12 @@ void BaseFilterWithMatrix::Apply(Image& image) {
                         actual_used_pixel_i = y_size - 1;
                     }
 
-                    new_pixel.red += image.pixels[actual_used_pixel_i][actual_used_pixel_j].red * matrix_[1 + shift_i][1 + shift_j] ;
-                    new_pixel.green += image.pixels[actual_used_pixel_i][actual_used_pixel_j].green * matrix_[1 + shift_i][1 + shift_j];
-                    new_pixel.blue += image.pixels[actual_used_pixel_i][actual_used_pixel_j].blue * matrix_[1 + shift_i][1 + shift_j];
+                    new_pixel.red +=
+                        image.pixels[actual_used_pixel_i][actual_used_pixel_j].red * matrix_[1 + shift_i][1 + shift_j];
+                    new_pixel.green += image.pixels[actual_used_pixel_i][actual_used_pixel_j].green *
+                                       matrix_[1 + shift_i][1 + shift_j];
+                    new_pixel.blue +=
+                        image.pixels[actual_used_pixel_i][actual_used_pixel_j].blue * matrix_[1 + shift_i][1 + shift_j];
                 }
             }
